+++ frontend/FamoNET/Model/Dto/DataSetDto.cs
@@ -5,9 +5,9 @@ namespace FamoNET.Model.Dto
     public class DataSetDto
     {
         [JsonPropertyName("mjd")]
-        public List<decimal> MJD { get; set; }
+        public List<double> MJD { get; set; }
 
         [JsonPropertyName("value")]
-        public List<decimal> Values { get; set; }
+        public List<double> Values { get; set; }
     }
 }
